requirements.txt
asgiref==3.3.0
certifi==2020.6.20
chardet==3.0.4
Django==3.0.10
django-braces==1.14.0
django-cors-headers==3.5.0
django-embed-video==1.3.2
django-environ==0.4.5
django-memcache-status==2.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.10
Pillow==7.0.0
PyJWT==1.7.1
python-dotenv==0.15.0
python-memcached==1.59
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
